==> dashboard/02_add_new_gpi_value.ipynb <==
import pandas as pd

#set the new year, e.g. 2021
new_year = 2021

Remeber later, that based on the differences of the countries found above replace the countries to be the same with the old dataframe with all the GPI values for all years. Below, we replace the differences found for the new year.

all_gpi = pd.read_csv('data/gpi_all_countries.csv') #read all GPI data

Download the .XLSX file for the Global Peace Index 2021 in the following adress:https://www.visionofhumanity.org/public-release-data/

#Keep only the new GPI values for each country so that to merge with the older data
gpioneyear = pd.read_excel('data/GPI_new.xlsx', sheet_name='Overall Scores', engine='openpyxl')

gpioneyear.columns = gpioneyear.iloc[2]
gpioneyear = gpioneyear[3:]

gpioneyear = gpioneyear[['Country', new_year]]

gpioneyear.columns = ['Country',str(int(a))]

Check the length of the files

len(all_gpi), len(gpioneyear)

Sort dataframes by country name

all_gpi = all_gpi.sort_values(by=['Country']).reset_index(drop = True)

gpioneyear = gpioneyear.sort_values(by=['Country']).reset_index(drop = True)

Find the differences for the country names between the older and newer GPI dataframe.

for el in list(all_gpi.Country): 
    if not(el in list(gpioneyear.Country)):
        print (el)

for el in list(gpioneyear.Country): 
    if not(el in list(all_gpi.Country)):
        print (el)

#Based on the differences of the countries found above replace the countries to be the same with the 
#old dataframe with all the GPI values for all years. Below, we replace the differences found for year 2021
gpioneyear = gpioneyear.replace({'Country':{'Burkina Faso' : 'BurkinaFaso', 'Cote d\' Ivoire' : 'Cote d’Ivoire',
                              'Czech Republic' : 'Czechia', 'Timor-Leste':'East Timor','The Gambia':'Gambia', 
                              'United States of America': 'United States'}})                                                                                         




Now that the country columns have no differences merge the data based on the Country columns

gpi_allyears = pd.merge(gpioneyear, all_gpi, on='Country')

gpi_allyears.to_csv('data/gpi_all_countries_new.csv', index=False)
