==> dashboard/03_Dataframe_GPI_score_fips_10_4_codes.ipynb <==
import pandas as pd

### Insert all GPI data from all countries as found online.

gpi = pd.read_csv('data/gpi_all_countries.csv')

gpi.head(3)

len(gpi)

Now we need to turn the countries in the dataset to "FIPS10-4" system in order to match it with the system used from GDELT.

Insert the "FIPS10-4" database.

Fips = pd.read_csv('data/Fips10_4.csv')

Fips.head(3)

There is usually a mess with the country names due to changes, etc. With the check and correct manually on the files until I have two documents with the same country names and then merge.

for el in sorted(list(gpi['Country'])): 
    if not(el in list(Fips['Country'])):
        print (el)

Merge to create the dataframe with all countries and years GPI score with the corresponding fips_10_4 codes.

gpifips = pd.merge(gpi, Fips, on='Country')
gpifips.head(1)

Get the columns in order to re-order them:

columnl = gpifips.columns.to_list()
columnl.reverse()

Reorder the columns in the dataframe:

col_list = [columnl[-1]] + [columnl[0]] + columnl[1:-1]

Ngpifips = gpifips[col_list]

Ngpifips.head(3)

len(Ngpifips)

Save the created dataframe:

Ngpifips.to_csv('data/Gpi_fips_10_4.csv',index=False)
